# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PINVEST BACKEND API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/faq/general": {
            "get": {
                "tags": [
                    "Faq"
                ],
                "operationId": "9d5c63d205809e9a70a04bb1836c28e8",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a listing of faq."
                    }
                }
            }
        },
        "/api/faq/{app}": {
            "get": {
                "tags": [
                    "Faq"
                ],
                "operationId": "251e21b6b80a3c86ac26e1cf8b2dcd0d",
                "parameters": [
                    {
                        "name": "app",
                        "in": "path",
                        "description": "app",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a listing of faq."
                    }
                }
            }
        },
...
    }
}